Refactoring unveil's code
Variables are now defined within the functions that use them in order
to help the browser do the garbage collection more efficiently.

The "attrib" variable was removed completely. The "unveil" event was
refactored as well and it is now more efficient so there is no need
to cache the default attribute to check for.

The filter method inside the "unveil" function is now more efficient
as well, as it doesn't use any jQuery methods anymore but the native
vanilla js equivalents.

The setting of the scroll and resize listeners is now done inline
using the "on" helper.

What's left?
Test if this refactored version works as well in older browsers as
the previous one did.